apps/backend/requirements.txt
fastapi>=0.112
uvicorn>=0.30
requests>=2.32
httpx>=0.27
python-dotenv>=1.0.1

SQLAlchemy>=2.0
psycopg2-binary>=2.9.9
pgvector>=0.2.5

pydantic>=2.7